Refactor code to have 2 separate packages for entities: model and business_rules
