+++ src/main/java/business_rules/PasswordEncryption.java
@@ -1,4 +1,4 @@
-package entity;
+package business_rules;
 
 public interface PasswordEncryption {
     String encryptMessage(String message);
