+++ src/main/java/model/ClothingItem.java
@@ -1,4 +1,4 @@
-package entity;
+package model;
 
 import java.awt.*;
 import java.util.Optional;
