+++ src/main/java/model/Location.java
@@ -1,4 +1,4 @@
-package entity;
+package model;
 
 public record Location(double longitude, double latitude) {
 }
