+++ src/main/java/model/Outfit.java
@@ -1,4 +1,4 @@
-package entity;
+package model;
 
 import java.util.Map;
 import java.util.Objects;
