+++ src/main/java/model/Weather.java
@@ -1,4 +1,4 @@
-package entity;
+package model;
 
 public record Weather(double temperature, boolean isRaining) {
 }
