+++ src/main/java/use_case/edit_wardrobe/update_clothing_item/UpdateDataAccess.java
@@ -1,6 +1,6 @@
 package use_case.edit_wardrobe.update_clothing_item;
 
-import entity.ClothingItem;
+import model.ClothingItem;
 
 public interface UpdateDataAccess {
     ClothingItem updateClothingItem(ClothingItem clothingItemToUpdate);
