+++ src/main/java/use_case/generate_outfit/OutfitOutputData.java
@@ -1,6 +1,6 @@
 package use_case.generate_outfit;
 
-import entity.Outfit;
+import model.Outfit;
 
 public record OutfitOutputData(Outfit outfit) {
 }
