+++ src/test/java/use_case/generate_outfit/OutfitGenerationInteractorTest.java
@@ -1,10 +1,8 @@
 package use_case.generate_outfit;
 
-import entity.ClothingItem;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.Collection;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
